+++ frontend/csi/controller_server.go
@@ -610,6 +610,10 @@ func (p *Plugin) CreateSnapshot(
 	if err != nil {
 		if utils.IsNotFoundError(err) {
 			return nil, status.Error(codes.NotFound, err.Error())
+		} else if utils.IsUnsupportedError(err) {
+			// CSI snapshotter has no exponential backoff for retries, so slow it down here
+			time.Sleep(10 * time.Second)
+			return nil, status.Error(codes.FailedPrecondition, err.Error())
 		}
 		return nil, status.Error(codes.Internal, err.Error())
 	}