+++ storage_drivers/gcp/gcp_cvs.go
@@ -1294,6 +1294,11 @@ func (d *NFSStorageDriver) Publish(
 	return nil
 }
 
+// CanSnapshot determines whether a snapshot as specified in the provided snapshot config may be taken.
+func (d *NFSStorageDriver) CanSnapshot(_ context.Context, _ *storage.SnapshotConfig) error {
+	return nil
+}
+
 // GetSnapshot gets a snapshot.  To distinguish between an API error reading the snapshot
 // and a non-existent snapshot, this method may return (nil, nil).
 func (d *NFSStorageDriver) GetSnapshot(