+++ storage_drivers/ontap/ontap_nas_flexgroup.go
@@ -729,6 +729,11 @@ func (d *NASFlexGroupStorageDriver) Publish(
 	return publishFlexVolShare(ctx, d.API, &d.Config, publishInfo, name)
 }
 
+// CanSnapshot determines whether a snapshot as specified in the provided snapshot config may be taken.
+func (d *NASFlexGroupStorageDriver) CanSnapshot(_ context.Context, _ *storage.SnapshotConfig) error {
+	return nil
+}
+
 // GetSnapshot gets a snapshot.  To distinguish between an API error reading the snapshot
 // and a non-existent snapshot, this method may return (nil, nil).
 func (d *NASFlexGroupStorageDriver) GetSnapshot(